move create_on_off to parameters module
