+++ qcodes/instrument_drivers/AimTTi/AimTTi_PL601P_channels.py
@@ -1,9 +1,8 @@
 from typing import Any, Dict, Optional
 
-from qcodes import ChannelList, InstrumentChannel, VisaInstrument
 from qcodes import validators as vals
-from qcodes.instrument.base import Instrument
-from qcodes.utils import create_on_off_val_mapping
+from qcodes.instrument import ChannelList, Instrument, InstrumentChannel, VisaInstrument
+from qcodes.parameters import create_on_off_val_mapping
 
 
 class NotKnownModel(Exception):
