+++ qcodes/instrument_drivers/Keysight/KtMAwg.py
@@ -2,9 +2,8 @@ import ctypes
 from functools import partial
 from typing import Any, Dict, Optional
 
-from qcodes.instrument.base import Instrument
-from qcodes.instrument.channel import InstrumentChannel
-from qcodes.utils import create_on_off_val_mapping
+from qcodes.instrument import Instrument, InstrumentChannel
+from qcodes.parameters import create_on_off_val_mapping
 from qcodes.validators import Numbers
 
 from .KtMAwgDefs import *
