+++ qcodes/instrument_drivers/Keysight/N9030B.py
@@ -3,8 +3,12 @@ from typing import Any, Dict, Tuple, Union
 import numpy as np
 
 from qcodes.instrument import InstrumentChannel, VisaInstrument
-from qcodes.parameters import Parameter, ParameterWithSetpoints, ParamRawDataType
-from qcodes.utils import create_on_off_val_mapping
+from qcodes.parameters import (
+    Parameter,
+    ParameterWithSetpoints,
+    ParamRawDataType,
+    create_on_off_val_mapping,
+)
 from qcodes.validators import Arrays, Enum, Ints, Numbers
