+++ qcodes/instrument_drivers/Keysight/keysight_e4980a.py
@@ -10,11 +10,9 @@ from qcodes.parameters import (
     ManualParameter,
     MultiParameter,
     ParamRawDataType,
-)
-from qcodes.utils import (
-    convert_legacy_version_to_supported_version,
     create_on_off_val_mapping,
 )
+from qcodes.utils import convert_legacy_version_to_supported_version
 from qcodes.validators import Bool, Enum, Ints, Numbers
