+++ qcodes/instrument_drivers/Keysight/keysightb1500/KeysightB1500_base.py
@@ -4,8 +4,7 @@ from collections import defaultdict
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
 
 from qcodes.instrument import VisaInstrument
-from qcodes.parameters import MultiParameter, Parameter
-from qcodes.utils import create_on_off_val_mapping
+from qcodes.parameters import MultiParameter, Parameter, create_on_off_val_mapping
 
 from . import constants
 from .KeysightB1500_module import (
