+++ qcodes/instrument_drivers/tektronix/Keithley_2450.py
@@ -5,8 +5,11 @@ import numpy as np
 from typing_extensions import TypedDict
 
 from qcodes.instrument import InstrumentChannel, VisaInstrument
-from qcodes.parameters import ParameterWithSetpoints, invert_val_mapping
-from qcodes.utils import create_on_off_val_mapping
+from qcodes.parameters import (
+    ParameterWithSetpoints,
+    create_on_off_val_mapping,
+    invert_val_mapping,
+)
 from qcodes.validators import Arrays, Enum, Ints, Lists, Numbers
