+++ qcodes/instrument_drivers/tektronix/keithley_7510.py
@@ -9,9 +9,9 @@ from qcodes.parameters import (
     MultiParameter,
     Parameter,
     ParamRawDataType,
+    create_on_off_val_mapping,
     invert_val_mapping,
 )
-from qcodes.utils import create_on_off_val_mapping
 from qcodes.validators import Arrays, Enum, Ints, Lists, Numbers
