+++ qcodes/tests/helpers/test_on_off_mapping.py
@@ -1,7 +1,6 @@
 import pytest
 
-from qcodes.parameters import invert_val_mapping
-from qcodes.utils import create_on_off_val_mapping
+from qcodes.parameters import create_on_off_val_mapping, invert_val_mapping
 
 
 def test_values_of_mapping_are_only_the_given_two():
